Module aws_sdk_cloudwatchevents::model

Modules

api_destination

See ApiDestination

archive

See Archive

aws_vpc_configuration

See AwsVpcConfiguration

batch_array_properties

See BatchArrayProperties

batch_parameters

See BatchParameters

batch_retry_strategy

See BatchRetryStrategy

capacity_provider_strategy_item

See CapacityProviderStrategyItem

condition

See Condition

connection

See Connection

connection_api_key_auth_response_parameters

See ConnectionApiKeyAuthResponseParameters

connection_auth_response_parameters

See ConnectionAuthResponseParameters

connection_basic_auth_response_parameters

See ConnectionBasicAuthResponseParameters

connection_body_parameter

See ConnectionBodyParameter

connection_header_parameter

See ConnectionHeaderParameter

connection_http_parameters

See ConnectionHttpParameters

connection_o_auth_client_response_parameters

See ConnectionOAuthClientResponseParameters

connection_o_auth_response_parameters

See ConnectionOAuthResponseParameters

connection_query_string_parameter

See ConnectionQueryStringParameter

create_connection_api_key_auth_request_parameters

See CreateConnectionApiKeyAuthRequestParameters

create_connection_auth_request_parameters

See CreateConnectionAuthRequestParameters

create_connection_basic_auth_request_parameters

See CreateConnectionBasicAuthRequestParameters

create_connection_o_auth_client_request_parameters

See CreateConnectionOAuthClientRequestParameters

create_connection_o_auth_request_parameters

See CreateConnectionOAuthRequestParameters

dead_letter_config

See DeadLetterConfig

ecs_parameters

See EcsParameters

event_bus

See EventBus

event_source

See EventSource

http_parameters

See HttpParameters

input_transformer

See InputTransformer

kinesis_parameters

See KinesisParameters

network_configuration

See NetworkConfiguration

partner_event_source

See PartnerEventSource

partner_event_source_account

See PartnerEventSourceAccount

placement_constraint

See PlacementConstraint

placement_strategy

See PlacementStrategy

put_events_request_entry

See PutEventsRequestEntry

put_events_result_entry

See PutEventsResultEntry

put_partner_events_request_entry

See PutPartnerEventsRequestEntry

put_partner_events_result_entry

See PutPartnerEventsResultEntry

put_targets_result_entry

See PutTargetsResultEntry

redshift_data_parameters

See RedshiftDataParameters

remove_targets_result_entry

See RemoveTargetsResultEntry

replay

See Replay

replay_destination

See ReplayDestination

retry_policy

See RetryPolicy

rule

See Rule

run_command_parameters

See RunCommandParameters

run_command_target

See RunCommandTarget

sage_maker_pipeline_parameter

See SageMakerPipelineParameter

sage_maker_pipeline_parameters

See SageMakerPipelineParameters

sqs_parameters

See SqsParameters

tag

See Tag

target

See Target

update_connection_api_key_auth_request_parameters

See UpdateConnectionApiKeyAuthRequestParameters

update_connection_auth_request_parameters

See UpdateConnectionAuthRequestParameters

update_connection_basic_auth_request_parameters

See UpdateConnectionBasicAuthRequestParameters

update_connection_o_auth_client_request_parameters

See UpdateConnectionOAuthClientRequestParameters

update_connection_o_auth_request_parameters

See UpdateConnectionOAuthRequestParameters

Structs

ApiDestination

Contains details about an API destination.

Archive

An Archive object that contains details about an archive.

AwsVpcConfiguration

This structure specifies the VPC subnets and security groups for the task, and whether a

BatchArrayProperties

The array properties for the submitted job, such as the size of the array. The array size

BatchParameters

The custom parameters to be used when the target is an Batch job.

BatchRetryStrategy

The retry strategy to use for failed jobs, if the target is an Batch job. If you

CapacityProviderStrategyItem

The details of a capacity provider strategy. To learn more, see CapacityProviderStrategyItem in the Amazon ECS API Reference.

Condition

A JSON string which you can use to limit the event bus permissions you are granting to

Connection

Contains information about a connection.

ConnectionApiKeyAuthResponseParameters

Contains the authorization parameters for the connection if API Key is specified as the

ConnectionAuthResponseParameters

Contains the authorization parameters to use for the connection.

ConnectionBasicAuthResponseParameters

Contains the authorization parameters for the connection if Basic is specified as the

ConnectionBodyParameter

Additional parameter included in the body. You can include up to 100 additional body

ConnectionHeaderParameter

Additional parameter included in the header. You can include up to 100 additional header

ConnectionHttpParameters

Contains additional parameters for the connection.

ConnectionOAuthClientResponseParameters

Contains the client response parameters for the connection when OAuth is specified as the

ConnectionOAuthResponseParameters

Contains the response parameters when OAuth is specified as the authorization type.

ConnectionQueryStringParameter

Additional query string parameter for the connection. You can include up to 100 additional

CreateConnectionApiKeyAuthRequestParameters

Contains the API key authorization parameters for the connection.

CreateConnectionAuthRequestParameters

Contains the authorization parameters for the connection.

CreateConnectionBasicAuthRequestParameters

Contains the Basic authorization parameters to use for the connection.

CreateConnectionOAuthClientRequestParameters

Contains the Basic authorization parameters to use for the connection.

CreateConnectionOAuthRequestParameters

Contains the OAuth authorization parameters to use for the connection.

DeadLetterConfig

A DeadLetterConfig object that contains information about a dead-letter queue

EcsParameters

The custom parameters to be used when the target is an Amazon ECS task.

EventBus

An event bus receives events from a source and routes them to rules associated with that

EventSource

A partner event source is created by an SaaS partner. If a customer creates a partner

HttpParameters

These are custom parameter to be used when the target is an API Gateway REST APIs or

InputTransformer

Contains the parameters needed for you to provide custom input to a target based on one or

KinesisParameters

This object enables you to specify a JSON path to extract from the event and use as the

NetworkConfiguration

This structure specifies the network configuration for an ECS task.

PartnerEventSource

A partner event source is created by an SaaS partner. If a customer creates a partner

PartnerEventSourceAccount

The Amazon Web Services account that a partner event source has been offered to.

PlacementConstraint

An object representing a constraint on task placement. To learn more, see Task Placement Constraints in the Amazon Elastic Container Service Developer

PlacementStrategy

The task placement strategy for a task or service. To learn more, see Task Placement Strategies in the Amazon Elastic Container Service Service Developer

PutEventsRequestEntry

Represents an event to be submitted.

PutEventsResultEntry

Represents an event that failed to be submitted.

PutPartnerEventsRequestEntry

The details about an event generated by an SaaS partner.

PutPartnerEventsResultEntry

Represents an event that a partner tried to generate, but failed.

PutTargetsResultEntry

Represents a target that failed to be added to a rule.

RedshiftDataParameters

These are custom parameters to be used when the target is a Amazon Redshift cluster to invoke the

RemoveTargetsResultEntry

Represents a target that failed to be removed from a rule.

Replay

A Replay object that contains details about a replay.

ReplayDestination

A ReplayDestination object that contains details about a replay.

RetryPolicy

A RetryPolicy object that includes information about the retry policy

Rule

Contains information about a rule in Amazon EventBridge.

RunCommandParameters

This parameter contains the criteria (either InstanceIds or a tag) used to specify which

RunCommandTarget

Information about the EC2 instances that are to be sent the command, specified as

SageMakerPipelineParameter

Name/Value pair of a parameter to start execution of a SageMaker Model Building

SageMakerPipelineParameters

These are custom parameters to use when the target is a SageMaker Model Building Pipeline

SqsParameters

This structure includes the custom parameter to be used when the target is an SQS FIFO

Tag

A key-value pair associated with an Amazon Web Services resource. In EventBridge, rules and event buses

Target

Targets are the resources to be invoked when a rule is triggered. For a complete list of

UpdateConnectionApiKeyAuthRequestParameters

Contains the API key authorization parameters to use to update the connection.

UpdateConnectionAuthRequestParameters

Contains the additional parameters to use for the connection.

UpdateConnectionBasicAuthRequestParameters

Contains the Basic authorization parameters for the connection.

UpdateConnectionOAuthClientRequestParameters

Contains the OAuth authorization parameters to use for the connection.

UpdateConnectionOAuthRequestParameters

Contains the OAuth request parameters to use for the connection.

Enums

ApiDestinationHttpMethod

ApiDestinationState

ArchiveState

AssignPublicIp

ConnectionAuthorizationType

ConnectionOAuthHttpMethod

ConnectionState

EventSourceState

LaunchType

PlacementConstraintType

PlacementStrategyType

PropagateTags

ReplayState

RuleState