Module aws_sdk_cloudwatchevents::operation

Structs

ActivateEventSource

Activates a partner event source that has been deactivated. Once activated, your matching

CancelReplay

Cancels the specified replay.

CreateApiDestination

Creates an API destination, which is an HTTP invocation endpoint configured as a target

CreateArchive

Creates an archive of events with the specified settings. When you create an archive,

CreateConnection

Creates a connection. A connection defines the authorization type and credentials to use

CreateEventBus

Creates a new event bus within your account. This can be a custom event bus which you can

CreatePartnerEventSource

Called by an SaaS partner to create a partner event source. This operation is not used by

DeactivateEventSource

You can use this operation to temporarily stop receiving events from the specified partner

DeauthorizeConnection

Removes all authorization parameters from the connection. This lets you remove the secret

DeleteApiDestination

Deletes the specified API destination.

DeleteArchive

Deletes the specified archive.

DeleteConnection

Deletes a connection.

DeleteEventBus

Deletes the specified custom event bus or partner event bus. All rules associated with

DeletePartnerEventSource

This operation is used by SaaS partners to delete a partner event source. This operation

DeleteRule

Deletes the specified rule.

DescribeApiDestination

Retrieves details about an API destination.

DescribeArchive

Retrieves details about an archive.

DescribeConnection

Retrieves details about a connection.

DescribeEventBus

Displays details about an event bus in your account. This can include the external Amazon Web Services

DescribeEventSource

This operation lists details about a partner event source that is shared with your

DescribePartnerEventSource

An SaaS partner can use this operation to list details about a partner event source that

DescribeReplay

Retrieves details about a replay. Use DescribeReplay to determine the

DescribeRule

Describes the specified rule.

DisableRule

Disables the specified rule. A disabled rule won't match any events, and won't

EnableRule

Enables the specified rule. If the rule does not exist, the operation fails.

ListApiDestinations

Retrieves a list of API destination in the account in the current Region.

ListArchives

Lists your archives. You can either list all the archives or you can provide a prefix to

ListConnections

Retrieves a list of connections from the account.

ListEventBuses

Lists all the event buses in your account, including the default event bus, custom event

ListEventSources

You can use this to see all the partner event sources that have been shared with your Amazon Web Services

ListPartnerEventSourceAccounts

An SaaS partner can use this operation to display the Amazon Web Services account ID that a particular

ListPartnerEventSources

An SaaS partner can use this operation to list all the partner event source names that

ListReplays

Lists your replays. You can either list all the replays or you can provide a prefix to

ListRuleNamesByTarget

Lists the rules for the specified target. You can see which of the rules in Amazon

ListRules

Lists your Amazon EventBridge rules. You can either list all the rules or you can provide

ListTagsForResource

Displays the tags associated with an EventBridge resource. In EventBridge, rules and event

ListTargetsByRule

Lists the targets assigned to the specified rule.

PutEvents

Sends custom events to Amazon EventBridge so that they can be matched to rules.

PutPartnerEvents

This is used by SaaS partners to write events to a customer's partner event bus. Amazon Web Services

PutPermission

Running PutPermission permits the specified Amazon Web Services account or Amazon Web Services organization

PutRule

Creates or updates the specified rule. Rules are enabled by default, or based on value of

PutTargets

Adds the specified targets to the specified rule, or updates the targets if they are

RemovePermission

Revokes the permission of another Amazon Web Services account to be able to put events to the specified

RemoveTargets

Removes the specified targets from the specified rule. When the rule is triggered, those

StartReplay

Starts the specified replay. Events are not necessarily replayed in the exact same order

TagResource

Assigns one or more tags (key-value pairs) to the specified EventBridge resource. Tags can

TestEventPattern

Tests whether the specified event pattern matches the provided event.

UntagResource

Removes one or more tags from the specified EventBridge resource. In Amazon EventBridge

UpdateApiDestination

Updates an API destination.

UpdateArchive

Updates the specified archive.

UpdateConnection

Updates settings for a connection.